Refactor ClientEventHandlers -> ClientTickEventHandlers for consistency, remove unused import from main class
